Task 4 Simple Calculator, #Oasisinfobyte
It takes a little skill to create a calculator with basic mathematical functionalities using CSS, HTML, and JavaScript. Calculators require an interactive interface to perform basic functions such as addition, subtraction, division, and multiplication. It will need a  display screen to display the user’s input and give the results. The grid system in CSS is for the alignments of buttons on the calculator. The additional tools you need to build a fully functional calculator include event Listeners, if-else statements, operators, loops, and so on.